refactor(home): extract NativeTilesSection into its own file
Shrinks add-tile-drawer.tsx by relocating the self-contained built-in
tiles section (no shared closures with the rest of the drawer) into its
own module.
